ServerSentEvents/line-contains-colon.html

<!DOCTYPE html>
<html>
<head>
	<title>Lines must be processed, in the order they are received, as follows:  If the line contains a U+003A COLON character (:), collect the characters on the line before the first (:), and let field be that string, collect the characters on the line after the first (:), and let value be that string. If value starts with a U+0020 SPACE character, remove it from value:</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>Lines must be processed, in the order they are received, as follows:  If the line contains a U+003A COLON character (:), collect the characters on the line before the first (:), and let field be that string, collect the characters on the line after the first (:), and let value be that string. If value starts with a U+0020 SPACE character, remove it from value:</h1>
	<div id="log"></div>
  <script>
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0072');
    t = async_test("If the line contains a U+003A COLON character (:), collect the characters on the line before the first (:), and let field be that string, collect the characters on the line after the first (:), and let value be that string. If value starts with a U+0020 SPACE character, remove it from value.");
    u = async_test("If value starts with more than one U+0020 SPACE character, remove one SPACE character from value.");
    s.onopen = function(e){ 
    	/* Connection opened. */
    };
    	
    s.onmessage = function(e){
    	var dat = e.data.split("\n");
    	console.log(dat);
    	t.step(function(){assert_equals(dat[1], "one space character")});
    	u.step(function(){assert_equals(dat[2], " more spaces character")});
    	t.done();
    	u.done();
    };
    	
    s.onerror = function(e){
    	/* Connection closed. */
    };

  </script>
  </div>
</body>
</html>
